@@ -23,7 +23,7 @@
     "@mdx-js/loader": "^3.0.1",
     "@mdx-js/react": "^3.0.1",
     "@metaplex-foundation/js": "^0.20.1",
-    "@next/mdx": "^14.2.4",
+    "@next/mdx": "15.1.4",
     "@radix-ui/react-checkbox": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-dropdown-menu": "^2.1.4",
@@ -49,13 +49,13 @@
     "jose": "^5.6.3",
     "lodash": "^4.17.21",
     "lucide-react": "^0.397.0",
-    "next": "^14.2.15",
+    "next": "15.1.4",
     "next-auth": "^4.24.7",
     "next-mdx-remote": "^5.0.0",
     "next-pwa": "^5.6.0",
-    "react": "^18",
+    "react": "19.0.0",
     "react-canvas-confetti": "^2.0.7",
-    "react-dom": "^18",
+    "react-dom": "19.0.0",
     "react-hook-form": "^7.52.1",
     "react-markdown": "^9.0.1",
     "react-slick": "^0.30.2",
@@ -74,13 +74,14 @@
     "@types/bs58": "^4.0.4",
     "@types/lodash": "^4.17.5",
     "@types/mdx": "^2.0.13",
+    "@types/next-pwa": "^5.6.9",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "19.0.3",
+    "@types/react-dom": "19.0.2",
     "@types/react-slick": "^0.23.13",
     "@typescript-eslint/eslint-plugin": "^6",
     "eslint": "^8",
-    "eslint-config-next": "14.2.4",
+    "eslint-config-next": "15.1.4",
     "eslint-config-prettier": "^9.1.0",
     "postcss": "^8",
     "prettier": "^3.3.2",
@@ -88,5 +89,9 @@
     "tailwind-scrollbar": "^3.1.0",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
+  },
+  "resolutions": {
+    "@types/react": "19.0.3",
+    "@types/react-dom": "19.0.2"
   }
 }
