@@ -1,6 +1,7 @@
 import { Metadata } from 'next'
 
-export async function generateMetadata({ params }: { params: { slug: string } }): Promise<Metadata> {
+export async function generateMetadata(props: { params: Promise<{ slug: string }> }): Promise<Metadata> {
+  const params = await props.params
   const metadataImagePath = `/api/metadata/comic/${params.slug}`
 
   return {
