@@ -5,12 +5,16 @@ import { ComicIssueSortTag } from '@/models/comicIssue/comicIssueParams'
 import { redirect, RedirectType } from 'next/navigation'
 
 type Props = {
-  params: {
+  params: Promise<{
     slug: string
-  }
+  }>
 }
 
-export default async function ComicReadPage({ params: { slug } }: Props) {
+export default async function ComicReadPage(props: Props) {
+  const params = await props.params
+
+  const { slug } = params
+
   const issues = await fetchComicIssues({
     skip: 0,
     take: 1,
