@@ -3,8 +3,8 @@ import { DiscoverPageWrapper } from '@/components/discover/DiscoverPageWrapper'
 import { CreatorGrid } from '@/components/discover/CreatorGrid'
 import { getAccessToken } from '@/app/lib/utils/auth'
 
-export default function DiscoverCreatorsPage() {
-  const accessToken = getAccessToken()
+export default async function DiscoverCreatorsPage() {
+  const accessToken = await getAccessToken()
   return (
     <DiscoverPageWrapper>
       <CreatorGrid accessToken={accessToken} />
