@@ -1,10 +1,11 @@
 import { fetchComicIssue } from '@/app/lib/api/comicIssue/queries'
-import { generateMetadataImage } from '@/utils/helpers'
+import { generateMetadataImage } from '@/utils/metadata'
 import { getStatelessCoverFromComicIssue } from '@/utils/covers'
 import { DefaultMetadataImage } from '@/components/metadata/DefaultImage'
 import { FallbackMetadataImage } from '@/components/metadata/FallbackImage'
 
-export async function GET(request: Request, { params }: { params: { id: string } }) {
+export async function GET(request: Request, props: { params: Promise<{ id: string }> }) {
+  const params = await props.params
   const url = new URL(request.url)
   const rarity = url.searchParams.get('rarity')
   const comicIssue = await fetchComicIssue({ id: params.id })
