@@ -6,7 +6,7 @@ import { cookies } from 'next/headers'
 // import { redirect, RedirectType } from 'next/navigation'
 
 export const logoutAction = async () => {
-  const initCookies = cookies()
+  const initCookies = await cookies()
   initCookies.delete(accessTokenKey)
   initCookies.delete(refreshTokenKey)
   // redirect(RoutePath.Login, RedirectType.replace)
