@@ -9,7 +9,7 @@ import { getAccessToken } from '../../utils/auth'
 const { WALLET, UPDATE } = WALLET_QUERY_KEYS
 
 export const updateWallet = async (address: string, request: UpdateWalletData): Promise<Nullable<Wallet>> => {
-  const accessToken = getAccessToken()
+  const accessToken = await getAccessToken()
   const response = await fetchWrapper<Wallet>({
     accessToken,
     path: `${WALLET}/${UPDATE}/${address}`,
