@@ -2,5 +2,5 @@ import { accessTokenKey } from '@/constants/general'
 import { cookies } from 'next/headers'
 import { isTokenValid } from './jwt'
 
-export const isAuthenticatedUser = (): boolean => isTokenValid(getAccessToken())
-export const getAccessToken = () => cookies().get(accessTokenKey)?.value ?? ''
+export const isAuthenticatedUser = async (): Promise<boolean> => isTokenValid(await getAccessToken())
+export const getAccessToken = async () => (await cookies()).get(accessTokenKey)?.value ?? ''
