@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { JSX } from 'react'
 import { cn } from '@/lib/utils'
 
 export type Variant = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'span'
