@@ -2,9 +2,9 @@ export type Nullable<T> = T | null
 
 export type CommonDialogProps = { open: boolean; toggleDialog: () => void }
 export type ComicIssuePageParams = {
-  params: {
+  params: Promise<{
     id: string
-  }
+  }>
 }
 
 export type ParsedFormError = { message: string; paths: Array<string | number> }
